requirements.txt
gradio==5.23.3
openai-whisper>=20240930
groq==0.28.0
python-dotenv==1.0.1
torch>=2.9.0
reportlab>=4.4.10